# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "poprako-server",
    "description": "Event-driven backend service for manga translation project management",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/assignments": {
      "get": {
        "tags": [
          "assignments"
        ],
        "summary": "`GET /api/v1/assignments` — list assignments by chapter or owner.",
        "description": "Lists assignments. Exactly one of `chapter_id` or `owner_id` is required; `role` optionally narrows by a single role bit. `incl` embeds related rows; dotted values imply their parent segments. Examples: `/api/v1/assignments?chapter_id=c_1&role=1&incl=chapter.comic.workset.team`, `/api/v1/assignments?owner_id=u_1&incl=user`.",
        "operationId": "list_infos",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "query",
            "description": "Chapter mode: list assignments on this chapter. Mutually exclusive with\n`owner_id`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner_id",
            "in": "query",
            "description": "Owner-user mode: list assignments owned by this user. Mutually exclusive\nwith `chapter_id`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "Single role-bit filter. Must be a singular valid role bit; composite\nvalues are rejected.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RoleField"
            }
          },
          {
            "name": "incl",
            "in": "query",
            "description": "Related rows to embed. Repeatable. Values: `user`, `chapter`,\n`chapter.comic`, `chapter.comic.workset`, `chapter.comic.workset.team`,\n`chapter.creator`, `chapter.comic.creator`. Dotted values imply their\nparent segments.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/AssignmentInclOpt"
              }
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignments listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_Vec_AssignmentInfoVal"
                }
              }
            }
          },
          "403": {
            "description": "No permission to list these assignments"
          },
          "422": {
            "description": "Exactly one of chapter_id or owner_id is required"
          }
        }
      }
    },
    "/api/v1/assignments/join": {
      "post": {
        "tags": [
          "assignments"
        ],
        "summary": "`POST /api/v1/assignments/join` — join a chapter assignment with roles.",
        "operationId": "join",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinChapterAssignmentParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Joined assignment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_AssignmentInfoVal"
                }
              }
            }
          },
          "403": {
            "description": "Role not assignable or no permission"
          },
          "404": {
            "description": "Chapter not found"
          }
        }
      }
    },
    "/api/v1/assignments/{assignment_id}": {
      "delete": {
        "tags": [
          "assignments"
        ],
        "summary": "`DELETE /api/v1/assignments/{assignment_id}` — delete an assignment.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "description": "Assignment ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Assignment deleted"
          },
          "403": {
            "description": "No permission to delete this assignment"
          },
          "404": {
            "description": "Assignment not found"
          }
        }
      }
    },
...
    "/api/v1/chapters/{chapter_id}/assignment-invitations": {
      "get": {
        "tags": [
          "assignment-invitations"
        ],
        "summary": "`GET /api/v1/chapters/{chapter_id}/assignment-invitations` — list invitations.",
        "description": "Lists a chapter's assignment invitations. `pending` filters by consumption state. Example: `/api/v1/chapters/{chapter_id}/assignment-invitations?pending=true&offset=0&limit=20`.",
        "operationId": "list_infos",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "description": "Chapter ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pending",
            "in": "query",
            "description": "When `Some(true)`, returns only unconsumed invitations;\n`Some(false)` returns only consumed ones; `None` returns all.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset (0-based).",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitations listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_Vec_AssignmentInvitationInfoVal"
                }
              }
            }
          },
          "403": {
            "description": "No permission to list invitations in this chapter"
          }
        }
      }
    },
    "/api/v1/chapters/{chapter_id}/assignments/{user_id}/roles": {
      "put": {
        "tags": [
          "assignments"
        ],
        "summary": "`PUT /api/v1/chapters/{chapter_id}/assignments/{user_id}/roles` — update roles.",
        "operationId": "update_roles",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "description": "Chapter ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "Assignee user ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssignmentRolesParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Assignment roles updated"
          },
          "403": {
            "description": "No permission to update this assignment"
          },
          "404": {
            "description": "Assignment not found"
          },
          "422": {
            "description": "Path ids do not match body ids"
          }
        }
      }
    },
    "/api/v1/chapters/{chapter_id}/pages": {
      "get": {
        "tags": [
          "pages"
        ],
        "summary": "`GET /api/v1/chapters/{chapter_id}/pages` — list pages in a chapter.",
        "operationId": "list_infos",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "description": "Chapter ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Zero-based offset for paginated results.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pages listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_Vec_PageInfoVal"
                }
              }
            }
          },
          "403": {
            "description": "No permission to list pages in this chapter"
          }
        }
      },
      "delete": {
        "tags": [
          "pages"
        ],
        "summary": "`DELETE /api/v1/chapters/{chapter_id}/pages` — delete all pages in a chapter.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "chapter_id",
            "in": "path",
            "description": "Chapter ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "All pages deleted"
          },
          "403": {
            "description": "No permission to delete pages in this chapter"
          }
        }
      }
    },
...
    "/api/v1/comics/{comic_id}/chapters": {
      "get": {
        "tags": [
          "chapters"
        ],
        "summary": "`GET /api/v1/comics/{comic_id}/chapters` — list chapters in a comic.",
        "description": "Lists chapters of a comic. `incl` embeds related rows; dotted values imply their parent segments. Example: `/api/v1/comics/{comic_id}/chapters?incl=comic.workset.team&incl=creator&offset=0&limit=20`.",
        "operationId": "list_infos",
        "parameters": [
          {
            "name": "comic_id",
            "in": "path",
            "description": "Comic ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "incl",
            "in": "query",
            "description": "Related rows to embed. Repeatable. Values: `comic`, `comic.workset`,\n`comic.workset.team`, `comic.creator`, `creator`. Dotted values imply\ntheir parent segments.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ChapterInclOpt"
              }
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset (0-based).",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chapters listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_Vec_ChapterInfoVal"
                }
              }
            }
          },
          "403": {
            "description": "No permission to list chapters in this comic"
          }
        }
      }
    },
...
    "/api/v1/member-invitations/{member_invitation_id}/roles": {
      "put": {
        "tags": [
          "member-invitations"
        ],
        "summary": "`PUT /api/v1/member-invitations/{member_invitation_id}/roles` — update invitation roles.",
        "operationId": "update_roles",
        "parameters": [
          {
            "name": "member_invitation_id",
            "in": "path",
            "description": "Invitation ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberInvitationRolesParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Invitation roles updated"
          },
          "403": {
            "description": "No permission to update this invitation"
          },
          "404": {
            "description": "Invitation not found"
          },
          "422": {
            "description": "Path id does not match body id"
          }
        }
      }
    },
    "/api/v1/members": {
      "get": {
        "tags": [
          "members"
        ],
        "summary": "`GET /api/v1/members` — list members by team or owner.",
        "description": "Lists members. Exactly one of `owner_id` or `team_id` is required. In `owner_id` mode, `role` and `fuzzy_nickname` must be omitted. In `team_id` mode, `fuzzy_nickname` and `role` are optional. `incl` embeds related rows. Examples: `/api/v1/members?team_id=t_1&fuzzy_nickname=al&role=1&incl=user`, `/api/v1/members?owner_id=u_1&incl=team`.",
        "operationId": "list_infos",
        "parameters": [
          {
            "name": "owner_id",
            "in": "query",
            "description": "Owner-user mode: list teams/memberships owned by this user. Mutually\nexclusive with `team_id`; when set, `role` and `fuzzy_nickname` must be\nomitted.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "description": "Team mode: list members of this team. Mutually exclusive with\n`owner_id`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fuzzy_nickname",
            "in": "query",
            "description": "Substring filter on member nickname (team mode only).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "description": "Single role-bit filter (team mode only). Must be a singular valid role\nbit; composite values are rejected.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RoleField"
            }
          },
          {
            "name": "incl",
            "in": "query",
            "description": "Related rows to embed. Repeatable. Values: `user`, `team`.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/MemberInclOpt"
              }
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Members listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_Vec_MemberInfoVal"
                }
              }
            }
          },
          "403": {
            "description": "No permission to list members in this team"
          },
          "422": {
            "description": "Exactly one of owner_id or team_id is required, or owner_id combined with role/fuzzy_nickname"
          }
        }
      },
      "post": {
        "tags": [
          "members"
        ],
        "summary": "`POST /api/v1/members` — create a member under a team.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Member created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_CreateMemberPayload"
                }
              }
            }
          },
          "403": {
            "description": "No permission to create members in this team"
          },
          "404": {
            "description": "User or team not found"
          },
          "409": {
            "description": "User is already a member"
          }
        }
      }
    },
...
    "/api/v1/members/{member_id}/roles": {
      "put": {
        "tags": [
          "members"
        ],
        "summary": "`PUT /api/v1/members/{member_id}/roles` — update a member's roles.",
        "operationId": "update_roles",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "description": "Member ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRolesParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Member roles updated"
          },
          "403": {
            "description": "No permission to update this member"
          },
          "404": {
            "description": "Member not found"
          },
          "422": {
            "description": "Path id does not match body id"
          }
        }
      }
    },
...
    "/api/v1/pages/{page_id}/units": {
      "get": {
        "tags": [
          "units"
        ],
        "summary": "`GET /api/v1/pages/{page_id}/units` — list units under a page.",
        "operationId": "list_infos",
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "description": "Page ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Zero-based offset for paginated results.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Units listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_ListPageUnitInfosPayload"
                }
              }
            }
          },
          "403": {
            "description": "No permission to list units in this page"
          },
          "404": {
            "description": "Page not found"
          }
        }
      }
    },
...
    "/api/v1/system-mails": {
      "get": {
        "tags": [
          "system-mails"
        ],
        "summary": "`GET /api/v1/system-mails` — list the current user's system mails.",
        "operationId": "list_infos",
        "parameters": [
          {
            "name": "read",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "System mails listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_Vec_SystemMailInfoVal"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
...
    "/api/v1/teams": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "`GET /api/v1/teams` — list teams with pagination.",
        "description": "Lists teams. Omit `user_id` to list all teams (super-admin only, otherwise `403`); supply `user_id` to list teams that user has joined. Examples: `/api/v1/teams?user_id=u_1&offset=0&limit=20`, `/api/v1/teams?offset=0&limit=20` (super-admin).",
        "operationId": "list_infos",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "description": "Filter to teams joined by this user. Omit to list all teams\n(super-admin only).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Teams listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_Vec_TeamInfoVal"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Listing all teams requires super-admin"
          }
        }
      },
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "`POST /api/v1/teams` — create a new team.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Team created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_TeamInfoVal"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Only super-admins can create teams"
          }
        }
      }
    },
...
    "/api/v1/teams/{team_id}/announcements": {
      "get": {
        "tags": [
          "announcements"
        ],
        "summary": "`GET /api/v1/teams/{team_id}/announcements` — list a team's announcements.",
        "description": "Lists a team's announcements. `incl` embeds related rows. Example: `/api/v1/teams/{team_id}/announcements?incl=user&offset=0&limit=20`.",
        "operationId": "list_infos",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "description": "Team ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "incl",
            "in": "query",
            "description": "Related rows to embed. Repeatable. Values: `user`.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/AnnouncementInclOpt"
              }
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset (0-based).",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Announcements listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpBody_Vec_AnnouncementInfoVal"
                }
              }
            }
          },
          "403": {
            "description": "No permission to list announcements in this team"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignmentInclOpt": {
        "type": "string",
        "description": "Incl opts for assignment info queries.\n\nEach opt embeds additional related data into the returned\n`AssignmentInfoVal`. Dotted opts implicitly pull in the segments before\nthe dot (e.g. `chapter.comic.workset.team` also embeds `chapter`,\n`chapter.comic`, and `chapter.comic.workset`).",
        "enum": [
          "user",
          "chapter",
          "chapter.comic",
          "chapter.comic.workset",
          "chapter.comic.workset.team",
          "chapter.creator",
          "chapter.comic.creator"
        ]
      },
...
      "ChapterInclOpt": {
        "type": "string",
        "description": "Incl opts for chapter info queries.\n\nEach opt embeds additional related data into the returned\n`ChapterInfoVal`. Dotted opts implicitly pull in the segments before the\ndot (e.g. `comic.workset.team` also embeds `comic` and `comic.workset`).",
        "enum": [
          "comic",
          "comic.workset",
          "comic.workset.team",
          "comic.creator",
          "creator"
        ]
      },
...
      "JoinChapterAssignmentParams": {
        "type": "object",
        "description": "Input parameters for a user joining a chapter as a worker via role\nselection.\n\nThe role mask must contain role bits that are valid for volunteer\nassignment; the use case layer validates this before applying.",
        "required": [
          "chapter_id",
          "roles"
        ],
        "properties": {
          "chapter_id": {
            "type": "string"
          },
          "roles": {
            "$ref": "#/components/schemas/RoleMask"
          }
        }
      },
...
      "RoleField": {
        "type": "integer",
        "format": "int32",
        "description": "A singular role permission flag represented as a bit position.\n\nEach role is a single bit value:\n\n| Value | Name          | Description             |\n|-------|---------------|-------------------------|\n| 1     | `RAW_PROVIDER` | Raw provider            |\n| 2     | `TRANSLATOR`   | Translator              |\n| 4     | `PROOFREADER`  | Proofreader             |\n| 8     | `TYPESETTER`   | Typesetter              |\n| 16    | `REDRAWER`     | Redrawer                |\n| 32    | `REVIEWER`     | Reviewer                |\n| 64    | `PUBLISHER`    | Publisher               |\n| 128   | `ADMIN`        | Admin                   |\n| 256   | `BOT`          | Bot                     |\n\nOnly a **single** valid bit is accepted; composite values are rejected.",
        "example": 2,
        "minimum": 0
      },
...
      "UpdateAssignmentRolesParams": {
        "type": "object",
        "description": "Input parameters for updating assignment roles.",
        "required": [
          "chapter_id",
          "user_id",
          "roles"
        ],
        "properties": {
          "chapter_id": {
            "type": "string"
          },
          "roles": {
            "$ref": "#/components/schemas/RoleMask"
          },
          "user_id": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}